move integration test into test manual (#188)
doesn't work with ci
